bird agora tem seu próprio arquivo também
as responsabilidades sobre o passáro que estava na cena foram passadas pra ele também.
