+++ constants.py
@@ -10,3 +10,5 @@ SCROLL_SPEED = 200
 GROUND_LINE = DEFAULT_WIN_SIZE[1] - 64
 
 SKY_COLOR = (11, 200, 215)
+
+OFFSET_X = 50
